% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv}
\alias{cv}
\title{Cross validation}
\usage{
cv(kappa_list, X_Euc)
}
\arguments{
\item{kappa_list}{a list of different kappa}

\item{X_Euc}{data in 3-dim Euclidean coordinates}
}
\value{
the optimal kappa
}
\description{
This function generates the optimal kappa based on Kullback-Leibler loss
}
\examples{
kappa_list <- c(10, 20)
X_Euc <- matrix(c(1, 1, 1, 0, 0, 0, 0, 0, 0), ncol = 3)
cv(kappa_list, X_Euc)

}
