% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density_estimation.R
\name{f_hat}
\alias{f_hat}
\title{KDE}
\usage{
f_hat(x, X_Euc, kappa)
}
\arguments{
\item{x}{new data points}

\item{X_Euc}{data in 3-dim Euclidean coordinates}

\item{kappa}{the smoothing parameter}
}
\value{
the estimated density at x
}
\description{
It is the KDE implementation when q = 3 and K(t) = exp(t)
}
\examples{
x <- t(matrix(c(1, 0, 0)))
X_Euc <- matrix(c(1, 1, 1, 0, 0, 0, 0, 0, 0), ncol = 3)
f_hat(x, X_Euc, kappa = 10)

}
