% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_plot}
\alias{generate_plot}
\title{Visualization}
\usage{
generate_plot(X_Euc, kappa)
}
\arguments{
\item{X_Euc}{data in 3-dim Euclidean coordinates}

\item{kappa}{the smoothing parameter}
}
\value{
visualization results
}
\description{
This function generates visualization results
}
\examples{
X_Euc <- matrix(c(1, 1, 1, 0, 0, 0, 0, 0, 0), ncol = 3)
generate_plot(X_Euc, kappa = 10)

}
