% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coordinates.R
\name{transform_coordinates}
\alias{transform_coordinates}
\title{Transform coordinates}
\usage{
transform_coordinates(point_list)
}
\arguments{
\item{point_list}{a series of data points in spherical coordinates}
}
\value{
data in 3-dim Euclidean coordinates
}
\description{
This function transforms spherical coordinates (longitude, latitude) to 3-dim Euclidean coordinates
}
\examples{
X <- t(matrix(180, 90))
X_Euc <- transform_coordinates(X)

}
